+++ tests/constraint_graph.cpp
@@ -26,6 +26,7 @@
 #include <constraint_graph.hpp>
 #include <disparity_graph.hpp>
 #include <image.hpp>
+#include <lowest_penalties.hpp>
 #include <pgm_io.hpp>
 
 BOOST_AUTO_TEST_SUITE(ConstraintGraphTest)
@@ -46,14 +47,14 @@ BOOST_AUTO_TEST_CASE(check_nodes_indexing)
     struct Image image{*pgm_io.get_image()};
 
     struct DisparityGraph disparity_graph{image, image, 3, 1, 1};
-    struct ConstraintGraph constraint_graph{disparity2constraint(disparity_graph, 1)};
-    BOOST_REQUIRE_EQUAL(constraint_graph.disparity_graph, &disparity_graph);
+    struct LowestPenalties lowest_penalties{disparity_graph};
+    struct ConstraintGraph constraint_graph{disparity_graph, lowest_penalties, 1};
 
-    BOOST_CHECK_EQUAL(node_index(*(constraint_graph.disparity_graph), {{0, 0}, 0}), 0);
-    BOOST_CHECK_EQUAL(node_index(*(constraint_graph.disparity_graph), {{0, 0}, 2}), 2);
-    BOOST_CHECK_EQUAL(node_index(*(constraint_graph.disparity_graph), {{0, 1}, 0}), 3);
-    BOOST_CHECK_EQUAL(node_index(*(constraint_graph.disparity_graph), {{0, 1}, 2}), 5);
-    BOOST_CHECK_EQUAL(node_index(*(constraint_graph.disparity_graph), {{1, 0}, 0}), 6);
+    BOOST_CHECK_EQUAL(node_index(constraint_graph.disparity_graph, {{0, 0}, 0}), 0);
+    BOOST_CHECK_EQUAL(node_index(constraint_graph.disparity_graph, {{0, 0}, 2}), 2);
+    BOOST_CHECK_EQUAL(node_index(constraint_graph.disparity_graph, {{0, 1}, 0}), 3);
+    BOOST_CHECK_EQUAL(node_index(constraint_graph.disparity_graph, {{0, 1}, 2}), 5);
+    BOOST_CHECK_EQUAL(node_index(constraint_graph.disparity_graph, {{1, 0}, 0}), 6);
 }
 
 BOOST_AUTO_TEST_CASE(check_black_images)
@@ -72,8 +73,8 @@ BOOST_AUTO_TEST_CASE(check_black_images)
     struct Image image{*pgm_io.get_image()};
 
     struct DisparityGraph disparity_graph{image, image, 3, 1, 1};
-    struct ConstraintGraph constraint_graph{disparity2constraint(disparity_graph, 1)};
-    BOOST_REQUIRE_EQUAL(constraint_graph.disparity_graph, &disparity_graph);
+    struct LowestPenalties lowest_penalties{disparity_graph};
+    struct ConstraintGraph constraint_graph{disparity_graph, lowest_penalties, 1};
     BOOST_CHECK_CLOSE(constraint_graph.threshold, 1, 1);
 
     struct Node node{{0, 0}, 0};
@@ -121,8 +122,8 @@ BOOST_AUTO_TEST_CASE(check_equal_images)
     struct Image image{*pgm_io.get_image()};
 
     struct DisparityGraph disparity_graph{image, image, 3, 1, 1};
-    struct ConstraintGraph constraint_graph{disparity2constraint(disparity_graph, 128 * 128)};
-    BOOST_REQUIRE_EQUAL(constraint_graph.disparity_graph, &disparity_graph);
+    struct LowestPenalties lowest_penalties{disparity_graph};
+    struct ConstraintGraph constraint_graph{disparity_graph, lowest_penalties, 128 * 128};
     BOOST_CHECK_CLOSE(constraint_graph.threshold, 128 * 128, 1);
 
     BOOST_CHECK_CLOSE(node_penalty(disparity_graph, {{0, 0}, 0}), 0, 1);
@@ -171,8 +172,8 @@ BOOST_AUTO_TEST_CASE(check_disparity_loop)
     struct Image right_image{*pgm_io.get_image()};
 
     struct DisparityGraph disparity_graph{left_image, right_image, 2, 1, 1};
-    struct ConstraintGraph constraint_graph{disparity2constraint(disparity_graph, 15)};
-    BOOST_REQUIRE_EQUAL(constraint_graph.disparity_graph, &disparity_graph);
+    struct LowestPenalties lowest_penalties{disparity_graph};
+    struct ConstraintGraph constraint_graph{disparity_graph, lowest_penalties, 15};
     BOOST_CHECK_CLOSE(constraint_graph.threshold, 15, 1);
 
     BOOST_CHECK_CLOSE(node_penalty(disparity_graph, {{0, 0}, 0}), 4, 1);
@@ -220,12 +221,70 @@ BOOST_AUTO_TEST_CASE(check_minimal_node_value_calculation)
     struct Image right_image{*pgm_io.get_image()};
 
     struct DisparityGraph disparity_graph{left_image, right_image, 3, 1, 1};
-    struct ConstraintGraph constraint_graph{disparity2constraint(disparity_graph, 0.5)};
-    BOOST_REQUIRE_EQUAL(constraint_graph.disparity_graph, &disparity_graph);
+    struct LowestPenalties lowest_penalties{disparity_graph};
+    struct ConstraintGraph constraint_graph{disparity_graph, lowest_penalties, 0.5};
     BOOST_CHECK_CLOSE(constraint_graph.threshold, 0.5, 1);
 
     BOOST_CHECK_CLOSE(node_penalty(disparity_graph, {{2, 0}, 0}), 1, 1);
     BOOST_CHECK(is_node_available(constraint_graph, {{2, 0}, 0}));
+}
+
+BOOST_AUTO_TEST_CASE(check_removal)
+{
+    PGM_IO pgm_io;
+    std::istringstream left_image_content{R"image(
+    P2
+    3 2
+    10
+    4 5 10
+    0 0 0
+    )image"};
+    std::istringstream right_image_content{R"image(
+    P2
+    3 2
+    10
+    10 7 0
+    0 0 0
+    )image"};
+
+    left_image_content >> pgm_io;
+    BOOST_REQUIRE(pgm_io.get_image());
+    struct Image left_image{*pgm_io.get_image()};
+
+    right_image_content >> pgm_io;
+    BOOST_REQUIRE(pgm_io.get_image());
+    struct Image right_image{*pgm_io.get_image()};
+
+    struct DisparityGraph disparity_graph{left_image, right_image, 3, 1, 1};
+    struct LowestPenalties lowest_penalties{disparity_graph};
+    struct ConstraintGraph constraint_graph{disparity_graph, lowest_penalties, 16};
+
+    BOOST_CHECK_CLOSE(constraint_graph.threshold, 16, 1);
+
+    BOOST_CHECK_CLOSE(node_penalty(disparity_graph, {{0, 0}, 0}), 36, 1);
+    BOOST_CHECK(!is_node_available(constraint_graph, {{0, 0}, 0}));
+
+    BOOST_CHECK_CLOSE(node_penalty(disparity_graph, {{0, 0}, 1}), 25, 1);
+    BOOST_CHECK(!is_node_available(constraint_graph, {{0, 0}, 1}));
+
+    BOOST_CHECK_CLOSE(node_penalty(disparity_graph, {{0, 0}, 2}), 0, 1);
+    BOOST_CHECK(is_node_available(constraint_graph, {{0, 0}, 2}));
+
+    BOOST_CHECK_CLOSE(node_penalty(disparity_graph, {{1, 0}, 0}), 4, 1);
+    BOOST_CHECK(is_node_available(constraint_graph, {{1, 0}, 0}));
+
+    BOOST_CHECK(solve_csp(&constraint_graph));
+
+    BOOST_CHECK(!is_node_available(constraint_graph, {{0, 0}, 0}));
+    BOOST_CHECK(!is_node_available(constraint_graph, {{0, 0}, 1}));
+    BOOST_CHECK(is_node_available(constraint_graph, {{0, 0}, 2}));
+
+    BOOST_CHECK(!is_node_available(constraint_graph, {{1, 0}, 0}));
+    BOOST_CHECK(is_node_available(constraint_graph, {{1, 0}, 1}));
+
+    BOOST_CHECK(!is_edge_available(constraint_graph, {{{0, 0}, 0}, {{1, 0}, 0}}));
+    BOOST_CHECK(!is_edge_available(constraint_graph, {{{0, 0}, 1}, {{1, 0}, 0}}));
+    BOOST_CHECK(!is_edge_available(constraint_graph, {{{0, 0}, 2}, {{1, 0}, 0}}));
 }
 
 BOOST_AUTO_TEST_SUITE_END()